+++ Source/Vulkan/VulkanVertex.hpp
@@ -32,8 +32,8 @@ public:
 		return vertices;
 	};
 
-	static std::vector<uint16_t> GetQuadVertexIndexCollection() {
-		std::vector<uint16_t> indices = {
+	static std::vector<uint32_t> GetQuadVertexIndexCollection() {
+		std::vector<uint32_t> indices = {
 			0, 1, 2, 2, 3, 0,
 			4, 5, 6, 6, 7, 4
 		};
